Proposal: expose MultiError (#8)
One of the great things about semgroup over x/sync/errgroup is that it
actually waits all tasks to complete and returns an error.

However, the returned error type doesn't let the caller to iterate over the
errors (e.g. my use case requires me to truncate the list of errors).

Therefore, introducing a backwards-compatible change that exports `MultiError`
type, and guarantees that `Wait()` method returns an error of this type.